Policy controls how the final value is saved into the archive
