+++ immer/extra/persist/json/policy.hpp
@@ -30,93 +30,29 @@ auto get_pools_types(const T&)
     return boost::hana::make_set();
 }
 
-struct via_get_pools_names_policy
+struct value0_serialize_t
 {
-    template <class T>
-    auto get_pool_types(const T& value) const
+    template <class Archive, class T>
+    void save(Archive& ar, const T& value0) const
     {
-        return boost::hana::keys(get_pools_names(value));
+        ar(CEREAL_NVP(value0));
     }
 
+    template <class Archive, class T>
+    void load(Archive& ar, T& value0) const
+    {
+        ar(CEREAL_NVP(value0));
+    }
+};
+
+struct no_wrap_t
+{
     auto get_output_wrap_fn() const { return boost::hana::id; }
     auto get_input_wrap_fn() const { return boost::hana::id; }
-
-    template <class T>
-    auto get_pool_name_fn(const T& value) const
-    {
-        using Map = decltype(get_pools_names(value));
-        return name_from_map_fn<Map>{};
-    }
 };
 
-struct via_get_pools_types_policy
+struct default_wrap_t
 {
-    template <class T>
-    auto get_pool_types(const T& value) const
-    {
-        return get_pools_types(value);
-    }
-
-    auto get_output_wrap_fn() const { return boost::hana::id; }
-    auto get_input_wrap_fn() const { return boost::hana::id; }
-
-    template <class T>
-    auto get_pool_name_fn(const T& value) const
-    {
-        return get_demangled_name_fn{};
-    }
-};
-
-struct hana_struct_auto_policy
-{
-    template <class T>
-    auto get_pool_types(const T& value) const
-    {
-        return get_pools_for_type(boost::hana::typeid_(value));
-    }
-
-    auto get_output_wrap_fn() const { return wrap_for_saving; }
-    auto get_input_wrap_fn() const { return wrap_for_loading; }
-
-    template <class T>
-    auto get_pool_name_fn(const T&) const
-    {
-        return get_demangled_name_fn{};
-    }
-};
-
-struct hana_struct_auto_member_name_policy
-{
-    template <class T>
-    auto get_pool_types(const T& value) const
-    {
-        return get_pools_for_type(boost::hana::typeid_(value));
-    }
-
-    auto get_output_wrap_fn() const { return wrap_for_saving; }
-    auto get_input_wrap_fn() const { return wrap_for_loading; }
-
-    template <class T>
-    auto get_pool_name_fn(const T& value) const
-    {
-        using map_t =
-            decltype(get_named_pools_for_type(boost::hana::typeid_(value)));
-        return name_from_map_fn<map_t>{};
-    }
-};
-
-template <class Map>
-struct via_map_policy
-{
-    static_assert(boost::hana::is_a<boost::hana::map_tag, Map>,
-                  "via_map_policy accepts a map of types to pool names");
-
-    template <class T>
-    auto get_pool_types(const T& value) const
-    {
-        return boost::hana::keys(Map{});
-    }
-
     auto get_output_wrap_fn() const
     {
         static_assert(
@@ -132,6 +68,90 @@ struct via_map_policy
         return wrap_for_saving;
     }
     auto get_input_wrap_fn() const { return wrap_for_loading; }
+};
+
+struct via_get_pools_names_policy
+    : value0_serialize_t
+    , no_wrap_t
+{
+    template <class T>
+    auto get_pool_types(const T& value) const
+    {
+        return boost::hana::keys(get_pools_names(value));
+    }
+
+    template <class T>
+    auto get_pool_name_fn(const T& value) const
+    {
+        using Map = decltype(get_pools_names(value));
+        return name_from_map_fn<Map>{};
+    }
+};
+
+struct demangled_names_t
+{
+    template <class T>
+    auto get_pool_name_fn(const T& value) const
+    {
+        return get_demangled_name_fn{};
+    }
+};
+
+struct via_get_pools_types_policy
+    : demangled_names_t
+    , value0_serialize_t
+    , no_wrap_t
+{
+    template <class T>
+    auto get_pool_types(const T& value) const
+    {
+        return get_pools_types(value);
+    }
+};
+
+struct hana_struct_auto_policy
+    : demangled_names_t
+    , default_wrap_t
+{
+    template <class T>
+    auto get_pool_types(const T& value) const
+    {
+        return get_pools_for_type(boost::hana::typeid_(value));
+    }
+};
+
+struct hana_struct_auto_member_name_policy
+    : value0_serialize_t
+    , default_wrap_t
+{
+    template <class T>
+    auto get_pool_types(const T& value) const
+    {
+        return get_pools_for_type(boost::hana::typeid_(value));
+    }
+
+    template <class T>
+    auto get_pool_name_fn(const T& value) const
+    {
+        using map_t =
+            decltype(get_named_pools_for_type(boost::hana::typeid_(value)));
+        return name_from_map_fn<map_t>{};
+    }
+};
+
+template <class Map>
+struct via_map_policy
+    : value0_serialize_t
+    , default_wrap_t
+{
+    static_assert(boost::hana::is_a<boost::hana::map_tag, Map>,
+                  "via_map_policy accepts a map of types to pool names");
+
+    template <class T>
+    auto get_pool_types(const T& value) const
+    {
+        return boost::hana::keys(Map{});
+    }
 
     template <class T>
     auto get_pool_name_fn(const T&) const
